CON-4938: fixed CR issues
