+++ Tests/Functional/Subscribers/LifecycleTest.php
@@ -30,6 +30,69 @@ class LifecycleTest extends \Enlight_Components_Test_Plugin_TestCase
         Shopware()->Plugins()->Backend()->Auth()->setNoAcl();
     }
 
+    public function testUpdatePrices()
+    {
+        $this->importFixtures(__DIR__ . '/_fixtures/simple_variants.sql');
+
+        $this->Request()
+            ->setMethod('POST')
+//            ->setPost('prices',
+//                [
+//                    0 => [
+//                        'id' => $price->getId(),
+//                        'from' => 1,
+//                        'to' => 5,
+//                        'price' => 238.00,
+//                        'pseudoPrice' => 0,
+//                        'percent' => 0,
+//                        'cloned' => false,
+//                        'customerGroupKey' => 'EK',
+//                        'customerGroup' => [
+//                            0 => [
+//                                'id' => 1,
+//                                'key' => 'EK',
+//                                'name' => 'Shopkunden',
+//                                'tax' => true,
+//                                'taxInput' => true,
+//                                'mode' => false,
+//                                'discount' => 0,
+//                            ],
+//                        ],
+//                    ],
+//                ])
+            ->setPost('controller', 'Article')
+            ->setPost('module', 'backend')
+            ->setPost('action', 'saveDetail')
+            ->setPost('number', 'sw32870.3')
+            ->setPost('price', 238.00)
+            ->setPost('additionalText', 'L / Schwarz')
+            ->setPost('supplierNumber', '')
+            ->setPost('active', false)
+            ->setPost('inStock', 15)
+            ->setPost('stockMin', 0)
+            ->setPost('weight', 0)
+            ->setPost('kind', 1)
+            ->setPost('position', 0)
+            ->setPost('shippingFree', false)
+            ->setPost('minPurchase', 1)
+            ->setPost('purchasePrice', 38.99)
+            ->setPost('articleId', 32870)
+            ->setPost('standard', false)
+            ->setPost('id', 2404537);
+
+
+        $this->dispatch('backend/Article/saveDetail');
+        $this->assertEquals(200, $this->Response()->getHttpResponseCode());
+        $this->assertTrue($this->View()->success);
+
+        $changes = $this->manager->getConnection()->query('SELECT * FROM sw_connect_change WHERE c_entity_id = ? ORDER BY c_revision', ['32870-2404537'])->fetchAll();
+        $this->assertCount(1, $changes);
+        $updateChange = $changes[0];
+        $this->assertEquals('update', $updateChange['c_operation']);
+        $product = unserialize($updateChange['c_payload']);
+        $this->assertEquals(200.00, $product->price);
+    }
+
     public function test_generate_delete_change_for_variants()
     {
         $this->importFixtures(__DIR__ . '/_fixtures/simple_variants.sql');
