+++ Tests/Integration/Components/ImportServiceTest.php
@@ -14,16 +14,15 @@ use ShopwarePlugins\Connect\Components\CategoryResolver\AutoCategoryResolver;
 use ShopwarePlugins\Connect\Components\ConfigFactory;
 use ShopwarePlugins\Connect\Components\ImportService;
 use ShopwarePlugins\Connect\Components\RandomStringGenerator;
-use ShopwarePlugins\Connect\Tests\ConnectTestHelperTrait;
 use ShopwarePlugins\Connect\Tests\DatabaseTestCaseTrait;
-use Tests\ShopwarePlugins\Connect\ConnectTestHelper;
+use ShopwarePlugins\Connect\Tests\ProductBuilderTrait;
 use Shopware\Connect\Gateway\PDO;
 use Shopware\CustomModels\Connect\ProductToRemoteCategoryRepository;
 
 class ImportServiceTest extends \PHPUnit_Framework_TestCase
 {
     use DatabaseTestCaseTrait;
-    use ConnectTestHelperTrait;
+    use ProductBuilderTrait;
 
     /**
      * @var \ShopwarePlugins\Connect\Components\ImportService
@@ -105,7 +104,7 @@ class ImportServiceTest extends \PHPUnit_Framework_TestCase
         // map buecher category to some local category
         $localCategory = $this->categoryRepository->find(6);
         /** @var \Shopware\CustomModels\Connect\RemoteCategory $remoteCategory */
-        $remoteCategory = $this->remoteCategoryRepository->findOneBy(['categoryKey' => '/deutsch/bücher']);
+        $remoteCategory = $this->remoteCategoryRepository->findOneBy(['categoryKey' => '/bücher']);
         $remoteCategory->addLocalCategory($localCategory);
         $this->manager->persist($remoteCategory);
         $this->manager->flush();
@@ -155,7 +154,7 @@ class ImportServiceTest extends \PHPUnit_Framework_TestCase
         $localCategory = $this->categoryRepository->find(6);
         $localCategory2 = $this->categoryRepository->find(8);
         /** @var \Shopware\CustomModels\Connect\RemoteCategory $remoteCategory */
-        $remoteCategory = $this->remoteCategoryRepository->findOneBy(['categoryKey' => '/deutsch/bücher']);
+        $remoteCategory = $this->remoteCategoryRepository->findOneBy(['categoryKey' => '/bücher']);
         $remoteCategory->addLocalCategory($localCategory);
         $this->manager->persist($remoteCategory);
         $this->manager->flush();
@@ -242,7 +241,7 @@ class ImportServiceTest extends \PHPUnit_Framework_TestCase
         $this->manager->persist($localCategory);
 
         /** @var \Shopware\CustomModels\Connect\RemoteCategory $remoteCategory */
-        $remoteCategory = $this->remoteCategoryRepository->findOneBy(['categoryKey' => '/deutsch/bücher']);
+        $remoteCategory = $this->remoteCategoryRepository->findOneBy(['categoryKey' => '/bücher']);
         $remoteCategory->addLocalCategory($localCategory);
         $this->manager->persist($remoteCategory);
         $this->manager->flush();
@@ -279,7 +278,7 @@ class ImportServiceTest extends \PHPUnit_Framework_TestCase
     {
         $localCategory = $this->categoryRepository->find(35);
         /** @var \Shopware\CustomModels\Connect\RemoteCategory $remoteCategory */
-        $remoteCategory = $this->remoteCategoryRepository->findOneBy(['categoryKey' => '/deutsch/bücher']);
+        $remoteCategory = $this->remoteCategoryRepository->findOneBy(['categoryKey' => '/bücher']);
 
         $this->importService->importRemoteCategory($localCategory->getId(), $remoteCategory->getCategoryKey(), $remoteCategory->getLabel());
